+++ flink-connector-bigquery/src/test/java/com/google/cloud/flink/bigquery/sink/serializer/BigQuerySchemaProviderTest.java
@@ -18,12 +18,14 @@ package com.google.cloud.flink.bigquery.sink.serializer;
 
 import com.google.api.services.bigquery.model.TableFieldSchema;
 import com.google.api.services.bigquery.model.TableSchema;
+import com.google.protobuf.DescriptorProtos;
 import com.google.protobuf.DescriptorProtos.FieldDescriptorProto;
 import com.google.protobuf.Descriptors;
 import com.google.protobuf.Descriptors.Descriptor;
 import com.google.protobuf.Descriptors.DescriptorValidationException;
 import com.google.protobuf.Descriptors.FieldDescriptor;
 import org.apache.avro.Schema;
+import org.apache.avro.generic.GenericRecord;
 import org.junit.Test;
 
 import java.util.Arrays;
@@ -61,7 +63,7 @@ public class BigQuerySchemaProviderTest {
     private final TableSchema tableSchema = new TableSchema().setFields(fields);
 
     @Test
-    public void testPrimitiveTypesConversion() throws DescriptorValidationException {
+    public void testPrimitiveTypesConversion()  {
 
         BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(tableSchema);
         Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
@@ -151,26 +153,7 @@ public class BigQuerySchemaProviderTest {
         return new Schema.Parser().parse(avroSchemaString);
     }
 
-    @Test
-    public void testAllPrimitiveSchemaConversion() throws DescriptorValidationException {
-
-        String fieldString =
-                " \"fields\": [\n"
-                        + "   {\"name\": \"name\", \"type\": \"string\"},\n"
-                        + "   {\"name\": \"number\", \"type\": \"long\"},\n"
-                        + "   {\"name\": \"quantity\", \"type\": \"int\"},\n"
-                        + "   {\"name\": \"fixed_field\", \"type\": {\"type\": \"fixed\", \"size\": 10,\"name\": \"hash\" }},\n"
-                        + "   {\"name\": \"price\", \"type\": \"float\"},\n"
-                        + "   {\"name\": \"double_field\", \"type\": \"double\"},\n"
-                        + "   {\"name\": \"boolean_field\", \"type\": \"boolean\"},\n"
-                        + "   {\"name\": \"enum_field\", \"type\": {\"type\":\"enum\", \"symbols\": [\"A\", \"B\", \"C\", \"D\"], \"name\": \"ALPHABET\"}},\n"
-                        + "   {\"name\": \"byte_field\", \"type\": \"bytes\"}\n"
-                        + " ]\n";
-
-        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
-        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
-        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
-
+    private void assertPrimitive(Descriptor descriptor) {
         assertThat(descriptor.findFieldByNumber(1).toProto())
                 .isEqualTo(
                         FieldDescriptorProto.newBuilder()
@@ -255,7 +238,30 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testAllLogicalSchemaConversion() throws DescriptorValidationException {
+    public void testAllPrimitiveSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"name\", \"type\": \"string\"},\n"
+                        + "   {\"name\": \"number\", \"type\": \"long\"},\n"
+                        + "   {\"name\": \"quantity\", \"type\": \"int\"},\n"
+                        + "   {\"name\": \"fixed_field\", \"type\": {\"type\": \"fixed\", \"size\": 10,\"name\": \"hash\" }},\n"
+                        + "   {\"name\": \"price\", \"type\": \"float\"},\n"
+                        + "   {\"name\": \"double_field\", \"type\": \"double\"},\n"
+                        + "   {\"name\": \"boolean_field\", \"type\": \"boolean\"},\n"
+                        + "   {\"name\": \"enum_field\", \"type\": {\"type\":\"enum\", \"symbols\": [\"A\", \"B\", \"C\", \"D\"], \"name\": \"ALPHABET\"}},\n"
+                        + "   {\"name\": \"byte_field\", \"type\": \"bytes\"}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        assertPrimitive(descriptor);
+    }
+
+    @Test
+    public void testAllLogicalSchemaConversion(){
 
         String fieldString =
                 " \"fields\": [\n"
@@ -367,7 +373,7 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testAllUnionLogicalSchemaConversion() throws DescriptorValidationException {
+    public void testAllUnionLogicalSchemaConversion()  {
 
         String fieldString =
                 " \"fields\": [\n"
@@ -479,7 +485,7 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testAllUnionPrimitiveSchemaConversion() throws DescriptorValidationException {
+    public void testAllUnionPrimitiveSchemaConversion()  {
 
         String fieldString =
                 " \"fields\": [\n"
@@ -582,7 +588,7 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testUnionInRecordSchemaConversation() throws DescriptorValidationException {
+    public void testUnionInRecordSchemaConversation()  {
 
         String fieldString =
                 " \"fields\": [\n"
@@ -627,7 +633,7 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testRecordInRecordSchemaConversion() throws DescriptorValidationException {
+    public void testRecordInRecordSchemaConversion()  {
 
         String fieldString =
                 " \"fields\": [\n"
@@ -673,7 +679,7 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testRecordOfLogicalTypeSchemaConversion() throws DescriptorValidationException {
+    public void testRecordOfLogicalTypeSchemaConversion()  {
 
         String fieldString =
                 " \"fields\": [\n"
@@ -824,7 +830,7 @@ public class BigQuerySchemaProviderTest {
     }
 
     @Test
-    public void testDefaultValueSchemaConversion() throws DescriptorValidationException {
+    public void testDefaultValueSchemaConversion()  {
         String fieldString =
                 " \"fields\": [\n"
                         + "{\"name\": \"long_with_default\", \"type\": [\"long\", \"null\"], \"default\": 100}"
@@ -843,4 +849,651 @@ public class BigQuerySchemaProviderTest {
         assertThat(fieldDescriptorProto.hasDefaultValue()).isTrue();
         assertThat(fieldDescriptorProto.getDefaultValue()).isEqualTo("100");
     }
+
+    @Test
+    public void testMapOfUnionTypeSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"map_of_union\", \"type\": {\"type\": \"map\", \"values\": [\"float\", \"null\"]}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("map_of_union");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(field.hasTypeName()).isTrue();
+        Descriptors.Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(field.getTypeName());
+        FieldDescriptorProto fieldDescriptor = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("key");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_STRING);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        fieldDescriptor = nestedDescriptor.findFieldByNumber(2).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("value");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(2);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_FLOAT);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_OPTIONAL);
+    }
+
+    @Test
+    public void testMapOfArraySchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"map_of_array\", \"type\": {\"type\": \"map\", \"values\": {\"type\": \"array\", \"items\": \"long\", \"name\": \"array_in_map\"}}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("map_of_array");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(field.hasTypeName()).isTrue();
+        Descriptors.Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(field.getTypeName());
+        FieldDescriptorProto fieldDescriptor = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("key");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_STRING);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        fieldDescriptor = nestedDescriptor.findFieldByNumber(2).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("value");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(2);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_INT64);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+    }
+
+    @Test
+    public void testMapInRecordSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"record_with_map\", "
+                        + "\"type\": {\"name\": \"actual_record\", \"type\": \"record\", \"fields\": [{\"name\": \"map_in_record\", \"type\": { \"type\": \"map\", \"values\": \"long\"}}]}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("record_with_map");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(field.hasTypeName()).isTrue();
+
+        Descriptors.Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(field.getTypeName());
+        FieldDescriptorProto fieldDescriptorProto = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptorProto.getName()).isEqualTo("map_in_record");
+        assertThat(fieldDescriptorProto.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptorProto.getType())
+                .isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(fieldDescriptorProto.getLabel())
+                .isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(fieldDescriptorProto.hasTypeName()).isTrue();
+
+        nestedDescriptor =
+                nestedDescriptor.findNestedTypeByName(fieldDescriptorProto.getTypeName());
+        fieldDescriptorProto = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptorProto.getName()).isEqualTo("key");
+        assertThat(fieldDescriptorProto.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptorProto.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_STRING);
+        assertThat(fieldDescriptorProto.getLabel())
+                .isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        fieldDescriptorProto = nestedDescriptor.findFieldByNumber(2).toProto();
+        assertThat(fieldDescriptorProto.getName()).isEqualTo("value");
+        assertThat(fieldDescriptorProto.getNumber()).isEqualTo(2);
+        assertThat(fieldDescriptorProto.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_INT64);
+        assertThat(fieldDescriptorProto.getLabel())
+                .isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+    }
+    @Test
+    public void testArrayOfUnionValueSchemaConversion()  {
+        String fieldString =
+                " \"fields\": [\n"
+                        + "{\"name\": \"array_with_union\", \"type\": "
+                        + "{\"type\": \"array\", \"items\":  [\"long\", \"null\"]}}"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        IllegalArgumentException exception =
+                assertThrows(
+                        IllegalArgumentException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception).hasMessageThat().contains("Array cannot have a NULLABLE element");
+    }
+
+    @Test
+    public void testNestedArraysSchemaConversion() {
+        String fieldString =
+                " \"fields\": [\n"
+                        + "{\"name\": \"nested_arrays\", \"type\":{\"type\": \"array\", \"items\": "
+                        + "{\"name\": \"array_inside\", \"type\": \"array\", \"items\": \"long\"}"
+                        + "}}"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+
+        IllegalStateException exception =
+                assertThrows(
+                        IllegalStateException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception).hasMessageThat().contains("Nested arrays not supported by BigQuery.");
+    }
+
+    @Test
+    public void testUnionOfMultipleDatatypeSchemaConversion() {
+        String fieldString =
+                " \"fields\": [\n"
+                        + "{\"name\": \"multiple_type_union\","
+                        + " \"type\":[\"null\", \"string\", \"int\"]"
+                        + "}"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        IllegalArgumentException exception =
+                assertThrows(
+                        IllegalArgumentException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception)
+                .hasMessageThat()
+                .contains("Multiple non-null union types are not supported.");
+    }
+
+    @Test
+    public void testRecordOfRecordSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"record_in_record\", \"type\": {\"name\": \"record_name\", \"type\": \"record\", \"fields\": "
+                        + "[{ \"name\":\"record_field\", \"type\": "
+                        + getRecord("record_inside_record")
+                        + "}]"
+                        + "}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("record_in_record");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(field.hasTypeName()).isTrue();
+        Descriptors.Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(field.getTypeName());
+
+        field = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("record_field");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(field.hasTypeName()).isTrue();
+
+        nestedDescriptor = nestedDescriptor.findNestedTypeByName(field.getTypeName());
+        field = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_INT64);
+        assertThat(field.getName()).isEqualTo("value");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+
+        field = nestedDescriptor.findFieldByNumber(2).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_STRING);
+        assertThat(field.getName()).isEqualTo("another_value");
+        assertThat(field.getNumber()).isEqualTo(2);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+    }
+
+    @Test
+    public void testMapOfMapSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"map_of_map\", \"type\": {\"type\": \"map\", \"values\": {\"type\": \"map\", \"values\": \"bytes\"}}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("map_of_map");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(field.hasTypeName()).isTrue();
+        Descriptors.Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(field.getTypeName());
+        FieldDescriptorProto fieldDescriptor = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("key");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_STRING);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        fieldDescriptor = nestedDescriptor.findFieldByNumber(2).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("value");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(2);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(fieldDescriptor.hasTypeName()).isTrue();
+        assertThat(nestedDescriptor.findNestedTypeByName(fieldDescriptor.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(fieldDescriptor.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_STRING)
+                                                .setName("key")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_BYTES)
+                                                .setName("value")
+                                                .setNumber(2)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .build());
+    }
+
+    @Test
+    public void testMapOfRecordSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"map_of_records\", \"type\": {\"type\": \"map\", \"values\": "
+                        + getRecord("record_inside_map")
+                        + "}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("map_of_records");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(field.hasTypeName()).isTrue();
+        Descriptors.Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(field.getTypeName());
+        FieldDescriptorProto fieldDescriptor = nestedDescriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("key");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_STRING);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        fieldDescriptor = nestedDescriptor.findFieldByNumber(2).toProto();
+        assertThat(fieldDescriptor.getName()).isEqualTo("value");
+        assertThat(fieldDescriptor.getNumber()).isEqualTo(2);
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(fieldDescriptor.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(fieldDescriptor.hasTypeName()).isTrue();
+        assertThat(nestedDescriptor.findNestedTypeByName(fieldDescriptor.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(fieldDescriptor.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_INT64)
+                                                .setName("value")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_STRING)
+                                                .setName("another_value")
+                                                .setNumber(2)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .build());
+    }
+
+    @Test
+    public void testRecordOfArraySchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"record_with_array\", \"type\": {\"name\": \"record_with_array_field\", \"type\": \"record\", \"fields\": [{\"name\": \"array_in_record\", \"type\": {\"type\": \"array\", \"items\": \"boolean\"}}]}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("record_with_array");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(field.hasTypeName()).isTrue();
+        assertThat(descriptor.findNestedTypeByName(field.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(field.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_BOOL)
+                                                .setName("array_in_record")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REPEATED)
+                                                .build())
+                                .build());
+    }
+
+    @Test
+    public void testArrayOfUnionOfMapSchemaConversion() {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"array_of_map_union\", \"type\": [\"null\", {\"type\": \"array\", \"items\": {\"type\": \"map\", \"values\": \"bytes\"}}]}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        UnsupportedOperationException exception =
+                assertThrows(
+                        UnsupportedOperationException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception)
+                .hasMessageThat()
+                .contains("MAP/ARRAYS in UNION types are not supported");
+    }
+
+    @Test
+    public void testArrayOfMapSchemaConversion() {
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"array_of_map\", \"type\": {\"type\": \"array\", \"items\": {\"type\": \"map\", \"values\": \"bytes\"}}}\n"
+                        + " ]\n";
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        UnsupportedOperationException exception =
+                assertThrows(
+                        UnsupportedOperationException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception).hasMessageThat().contains("Array of Type MAP not supported yet.");
+    }
+
+    @Test
+    public void testArrayOfRecordSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "{\"name\": \"array_of_records\", \"type\":{\"type\": \"array\", \"items\": "
+                        + getRecord("inside_record")
+                        + "}}"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("array_of_records");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(field.hasTypeName()).isTrue();
+        assertThat(descriptor.findNestedTypeByName(field.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(field.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_INT64)
+                                                .setName("value")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_STRING)
+                                                .setName("another_value")
+                                                .setNumber(2)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .build());
+    }
+
+    @Test
+    public void testUnionOfArrayOfRecordSchemaConversion() {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"array_of_records_union\", \"type\": [\"null\", {\"type\": \"array\", \"items\": "
+                        + getRecord("inside_record_union")
+                        + "}]}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+
+        UnsupportedOperationException exception =
+                assertThrows(
+                        UnsupportedOperationException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception)
+                .hasMessageThat()
+                .contains("MAP/ARRAYS in UNION types are not supported");
+    }
+
+    @Test
+    public void testUnionOfRecordSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "{\"name\": \"record_field_union\","
+                        + " \"type\": [\"null\", "
+                        + getRecord("inside_record")
+                        + "]}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("record_field_union");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_OPTIONAL);
+        assertThat(field.hasTypeName()).isTrue();
+        assertThat(descriptor.findNestedTypeByName(field.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(field.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_INT64)
+                                                .setName("value")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_STRING)
+                                                .setName("another_value")
+                                                .setNumber(2)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .build());
+    }
+
+    @Test
+    public void testSpecialSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"record_field\", \"type\": "
+                        + getRecord("inside_record")
+                        + "},\n"
+                        + "   {\"name\": \"map_field\", \"type\": {\"type\": \"map\", \"values\": \"long\"}},\n"
+                        + "   {\"name\": \"array_field\", \"type\": {\"type\": \"array\", \"items\": \"float\"}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto field = descriptor.findFieldByNumber(1).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("record_field");
+        assertThat(field.getNumber()).isEqualTo(1);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(field.hasTypeName()).isTrue();
+        assertThat(descriptor.findNestedTypeByName(field.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(field.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_INT64)
+                                                .setName("value")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_STRING)
+                                                .setName("another_value")
+                                                .setNumber(2)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .build());
+
+        field = descriptor.findFieldByNumber(2).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(field.getName()).isEqualTo("map_field");
+        assertThat(field.getNumber()).isEqualTo(2);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+        assertThat(field.hasTypeName()).isTrue();
+        assertThat(descriptor.findNestedTypeByName(field.getTypeName()).toProto())
+                .isEqualTo(
+                        DescriptorProtos.DescriptorProto.newBuilder()
+                                .setName(field.getTypeName())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_STRING)
+                                                .setName("key")
+                                                .setNumber(1)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .addField(
+                                        FieldDescriptorProto.newBuilder()
+                                                .setType(FieldDescriptorProto.Type.TYPE_INT64)
+                                                .setName("value")
+                                                .setNumber(2)
+                                                .setLabel(FieldDescriptorProto.Label.LABEL_REQUIRED)
+                                                .build())
+                                .build());
+
+        field = descriptor.findFieldByNumber(3).toProto();
+        assertThat(field.getType()).isEqualTo(FieldDescriptorProto.Type.TYPE_FLOAT);
+        assertThat(field.getName()).isEqualTo("array_field");
+        assertThat(field.getNumber()).isEqualTo(3);
+        assertThat(field.getLabel()).isEqualTo(FieldDescriptorProto.Label.LABEL_REPEATED);
+    }
+
+    @Test
+    public void testUnionOfArraySchemaConversion() {
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"array_field_union\", \"type\": [\"null\", {\"type\": \"array\", \"items\": \"float\"}]}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+
+        UnsupportedOperationException exception =
+                assertThrows(
+                        UnsupportedOperationException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception)
+                .hasMessageThat()
+                .contains("MAP/ARRAYS in UNION types are not supported");
+    }
+
+    @Test
+    public void testUnionofMapSchemaConversion() {
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"map_field_union\", \"type\": [\"null\", {\"type\": \"map\", \"values\": \"long\"}]}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+
+        UnsupportedOperationException exception =
+                assertThrows(
+                        UnsupportedOperationException.class,
+                        () -> new BigQuerySchemaProvider(avroSchema));
+        assertThat(exception)
+                .hasMessageThat()
+                .contains("MAP/ARRAYS in UNION types are not supported");
+    }
+
+    @Test
+    public void testAllPrimitiveSingleUnionSchemaConversion()  {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"name\", \"type\": [\"string\"]},\n"
+                        + "   {\"name\": \"number\", \"type\": [\"long\"]},\n"
+                        + "   {\"name\": \"quantity\", \"type\": [\"int\"]},\n"
+                        + "   {\"name\": \"fixed_field\", \"type\": [{\"type\": \"fixed\", \"size\": 10,\"name\": \"hash\" }]},\n"
+                        + "   {\"name\": \"price\", \"type\": [\"float\"]},\n"
+                        + "   {\"name\": \"double_field\", \"type\": [\"double\"]},\n"
+                        + "   {\"name\": \"boolean_field\", \"type\": [\"boolean\"]},\n"
+                        + "   {\"name\": \"enum_field\", \"type\": [{\"type\":\"enum\", \"symbols\": [\"A\", \"B\", \"C\", \"D\"], \"name\": \"ALPHABET\"}]},\n"
+                        + "   {\"name\": \"byte_field\", \"type\": [\"bytes\"]}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+        assertPrimitive(descriptor);
+    }
+    @Test
+    public void testRecordOfUnionFieldSchemaConversion() throws DescriptorValidationException {
+
+        String fieldString =
+                " \"fields\": [\n"
+                        + "   {\"name\": \"record_with_union\", \"type\": {\"name\": \"record_with_union_field\", \"type\": \"record\", \"fields\": [{\"name\": \"union_in_record\", \"type\": [\"boolean\", \"null\"], \"default\": true}]}}\n"
+                        + " ]\n";
+
+        Schema avroSchema = getAvroSchemaFromFieldString(fieldString);
+        BigQuerySchemaProvider bigQuerySchemaProvider = new BigQuerySchemaProvider(avroSchema);
+        Descriptor descriptor = bigQuerySchemaProvider.getDescriptor();
+
+        FieldDescriptorProto fieldDescriptorProto = descriptor.findFieldByNumber(1).toProto();
+        assertThat(fieldDescriptorProto.getName()).isEqualTo("record_with_union");
+        assertThat(fieldDescriptorProto.getNumber()).isEqualTo(1);
+        assertThat(fieldDescriptorProto.getLabel())
+                .isEqualTo(FieldDescriptorProto.Label.LABEL_REQUIRED);
+        assertThat(fieldDescriptorProto.getType())
+                .isEqualTo(FieldDescriptorProto.Type.TYPE_MESSAGE);
+        assertThat(fieldDescriptorProto.hasTypeName()).isTrue();
+
+        Descriptor nestedDescriptor =
+                descriptor.findNestedTypeByName(fieldDescriptorProto.getTypeName());
+        FieldDescriptor fieldDescriptor = nestedDescriptor.findFieldByNumber(1);
+        assertThat(fieldDescriptor.isOptional()).isTrue();
+        assertThat(fieldDescriptor.getType()).isEqualTo(FieldDescriptor.Type.BOOL);
+        assertThat(fieldDescriptor.getName()).isEqualTo("union_in_record");
+        assertThat(fieldDescriptor.hasDefaultValue()).isTrue();
+        assertThat(fieldDescriptor.getDefaultValue()).isEqualTo(true);
+    }
+
 }
